ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
pillow==10.2.0
pytesseract==0.3.10
openai==1.10.0
anthropic==0.8.1
python-dotenv==1.0.0
